Move fixtures to test/experimental/conftest.py
