+++ tests/experimental/test_experimental_gaussian_process_exact.py
@@ -1,37 +1,10 @@
 import gpytorch
-import pytest
-import torch
 from autoemulate.emulators.gaussian_process import constant_mean, rbf, rbf_times_linear
 from autoemulate.experimental.emulators.gaussian_process.exact import (
     GaussianProcessExact,
 )
 from autoemulate.experimental.tuner import Tuner
 from autoemulate.experimental.types import DistributionLike
-from sklearn.datasets import make_regression
-
-
-@pytest.fixture
-def sample_data_y1d():
-    x, y = make_regression(n_samples=20, n_features=5, n_targets=1, random_state=0)  # type: ignore noqa: PGH003
-    return torch.Tensor(x), torch.Tensor(y)
-
-
-@pytest.fixture
-def new_data_y1d():
-    x, y = make_regression(n_samples=20, n_features=5, n_targets=1, random_state=1)  # type: ignore noqa: PGH003
-    return torch.Tensor(x), torch.Tensor(y)
-
-
-@pytest.fixture
-def sample_data_y2d():
-    x, y = make_regression(n_samples=20, n_features=5, n_targets=2, random_state=0)  # type: ignore noqa: PGH003
-    return torch.Tensor(x), torch.Tensor(y)
-
-
-@pytest.fixture
-def new_data_y2d():
-    x, y = make_regression(n_samples=20, n_features=5, n_targets=2, random_state=1)  # type: ignore noqa: PGH003
-    return torch.Tensor(x), torch.Tensor(y)
 
 
 def test_predict_with_uncertainty_gp(sample_data_y1d, new_data_y1d):
